+++ CHANGELOG.md
@@ -42,6 +42,7 @@
  * Added clean-up definition plugin in order to remove non used definitions of each endpoint
  * **[BC BREAK]** Moved plugins configuration out of `definitions`
  * Added option to set custom favicon in the GraphiQL Explorer
+ * Added option to set custom API documentation link in the GraphiQL Explorer
  
 1.0.6 - 18-03-23
 ----